+++ packages/react/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @components-design-ui/react
 
+## 3.0.0
+
+### Major Changes
+
+- Add components toast and tooltip
+
 ## 2.0.0
 
 ### Major Changes