+++ packages/tokens/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @components-design-ui/tokens
 
+## 2.0.0
+
+### Major Changes
+
+- Add components toast and tooltip
+
 ## 1.1.0
 
 ### Minor Changes